Refactor SDF module into organized submodules
- Split monolithic sdf.rs into focused submodules
- Add grid.rs for SDF grid data structures
- Add traits.rs for common SDF interfaces
- Add serial.rs and parallel.rs for different computation strategies
- Update readme.md with new module structure
